Replace legacy runner catalog by wanda catalog usage
