+++ assets/js/components/ChecksCatalog/ChecksCatalog.jsx
@@ -1,72 +1,70 @@
+/* eslint-disable react/no-array-index-key */
 import React, { useState, useEffect } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
 
-import { Disclosure, Transition } from '@headlessui/react';
+import { groupBy } from '@lib/lists';
 
-import ReactMarkdown from 'react-markdown';
-import remarkGfm from 'remark-gfm';
-
-import NotificationBox from '@components/NotificationBox';
-import LoadingBox from '@components/LoadingBox';
-
-import { EOS_ERROR } from 'eos-icons-react';
+import { getCatalog } from '@state/selectors/catalog';
+import { updateCatalog } from '@state/actions/catalog';
+import {
+  providerData,
+  getLabels,
+  getProviderByLabel,
+} from '@components/ProviderLabel/ProviderLabel';
+import CatalogContainer from './CatalogContainer';
+import CheckItem from './CheckItem';
 import ProviderSelection from './ProviderSelection';
 
+const ALL_FILTER = 'All';
+const updatedProvider = {
+  default: { label: ALL_FILTER },
+  ...providerData,
+};
+const providerLabels = getLabels(updatedProvider);
+
+// eslint-disable-next-line import/prefer-default-export
 function ChecksCatalog() {
   const dispatch = useDispatch();
+  const [selectedProvider, setProviderSelected] = useState(ALL_FILTER);
 
-  const [catalogData, catalogError, loading] = useSelector((state) => [
-    state.catalog.data,
-    state.catalog.error,
-    state.catalog.loading,
-  ]);
-
-  const providers = catalogData.map((provider) => provider.provider);
-  const [selected, setSelected] = useState(providers[0]);
-
-  const dispatchUpdateCatalog = () => {
-    dispatch({
-      type: 'UPDATE_CATALOG',
-      payload: {},
-    });
-  };
+  const {
+    data: catalogData,
+    error: catalogError,
+    loading,
+  } = useSelector(getCatalog());
 
   useEffect(() => {
-    setSelected(providers[0]);
-  }, [providers[0]]);
+    const apiParams =
+      selectedProvider === ALL_FILTER
+        ? {}
+        : { provider: getProviderByLabel(updatedProvider, selectedProvider) };
 
-  useEffect(() => {
-    dispatchUpdateCatalog();
-  }, [dispatch]);
-
-  if (loading) {
-    return <LoadingBox text="Loading checks catalog..." />;
-  }
-
-  if (catalogError) {
-    return (
-      <NotificationBox
-        icon={<EOS_ERROR className="m-auto" color="red" size="xl" />}
-        text={catalogError}
-        buttonText="Try again"
-        buttonOnClick={dispatchUpdateCatalog}
-      />
-    );
-  }
-
+    dispatch(updateCatalog(apiParams));
+  }, [dispatch, selectedProvider]);
   return (
-    <div>
+    <CatalogContainer
+      onRefresh={() =>
+        dispatch(
+          updateCatalog({
+            provider:
+              getProviderByLabel(providerData, selectedProvider) || null,
+          })
+        )
+      }
+      isCatalogEmpty={catalogData.length === 0}
+      catalogError={catalogError}
+      loading={loading}
+    >
       <ProviderSelection
-        providers={providers}
-        selected={selected}
-        onChange={setSelected}
+        providers={providerLabels}
+        selected={selectedProvider}
+        onChange={setProviderSelected}
       />
-      {catalogData
-        .filter((provider) => provider.provider === selected)
-        .map(({ _, groups }) =>
-          groups?.map(({ group, checks }) => (
+      <div>
+        {Object.entries(groupBy(catalogData, 'group')).map(
+          ([group, checks], idx) => (
             <div
-              key={group}
+              key={idx}
               className="check-group bg-white shadow overflow-hidden sm:rounded-md mb-8"
             >
               <div className="bg-white px-4 py-5 border-b border-gray-200 sm:px-6">
@@ -76,64 +74,20 @@ function ChecksCatalog() {
               </div>
               <ul className="divide-y divide-gray-200">
                 {checks.map((check) => (
-                  <li key={check.id}>
-                    <Disclosure>
-                      <Disclosure.Button
-                        as="div"
-                        className="flex justify-between w-full cursor-pointer hover:bg-gray-100"
-                      >
-                        <div className="check-row px-4 py-4 sm:px-6">
-                          <div className="flex items-center">
-                            <p className="px-2 inline-flex text-xs leading-5 font-semibold rounded-full bg-green-100 text-green-800">
-                              {check.id}
-                            </p>
-                            {check.premium > 0 && (
-                              <p className="px-2 inline-flex text-xs leading-5 font-semibold rounded-full bg-green-100 text-green-800">
-                                Premium
-                              </p>
-                            )}
-                          </div>
-                          <div className="mt-2 sm:flex sm:justify-between">
-                            <div className="sm:flex">
-                              <ReactMarkdown
-                                className="markdown text-sm"
-                                remarkPlugins={[remarkGfm]}
-                              >
-                                {check.description}
-                              </ReactMarkdown>
-                            </div>
-                          </div>
-                        </div>
-                      </Disclosure.Button>
-                      <Transition
-                        enter="transition duration-100 ease-out"
-                        enterFrom="transform opacity-0"
-                        enterTo="transform opacity-100"
-                        leave="transition duration-100 ease-out"
-                        leaveFrom="transform opacity-100"
-                        leaveTo="transform opacity-0"
-                      >
-                        <Disclosure.Panel className="check-panel border-none">
-                          <div className="px-8 py-4 sm:px-8">
-                            <div className="px-4 py-4 sm:px-4 bg-slate-100 rounded">
-                              <ReactMarkdown
-                                className="markdown"
-                                remarkPlugins={[remarkGfm]}
-                              >
-                                {check.remediation}
-                              </ReactMarkdown>
-                            </div>
-                          </div>
-                        </Disclosure.Panel>
-                      </Transition>
-                    </Disclosure>
-                  </li>
+                  <CheckItem
+                    key={check.id}
+                    checkID={check.id}
+                    premium={check.premium}
+                    description={check.description}
+                    remediation={check.remediation}
+                  />
                 ))}
               </ul>
             </div>
-          ))
+          )
         )}
-    </div>
+      </div>
+    </CatalogContainer>
   );
 }
