+++ assets/js/components/ChecksCatalog/index.js
@@ -1,5 +1,3 @@
 import ChecksCatalog from './ChecksCatalog';
 
-export { ChecksCatalogNew } from './ChecksCatalogNew';
-
 export default ChecksCatalog;
