+++ assets/js/state/actions/catalog.js
@@ -1,4 +1,4 @@
-export const UPDATE_CATALOG = 'UPDATE_CATALOG_NEW';
+export const UPDATE_CATALOG = 'UPDATE_CATALOG';
 
 export const updateCatalog = (env = {}) => ({
   type: UPDATE_CATALOG,
