+++ assets/js/trento.jsx
@@ -23,7 +23,7 @@ import { ExecutionResultsPage } from '@components/ExecutionResults';
 import SapSystemsOverview from '@components/SapSystemsOverview';
 import HostDetails from '@components/HostDetails';
 import DatabasesOverview from '@components/DatabasesOverview';
-import ChecksCatalog, { ChecksCatalogNew } from '@components/ChecksCatalog';
+import ChecksCatalog from '@components/ChecksCatalog';
 import NotFound from '@components/NotFound';
 import SomethingWentWrong from '@components/SomethingWentWrong';
 import Settings from '@components/Settings';
@@ -77,7 +77,6 @@ function App() {
                   <Route path="sap_systems" element={<SapSystemsOverview />} />
                   <Route path="databases" element={<DatabasesOverview />} />
                   <Route path="catalog" element={<ChecksCatalog />} />
-                  <Route path="catalog_new" element={<ChecksCatalogNew />} />
                   <Route path="settings" element={<Settings />} />
                   <Route path="about" element={<AboutPage />} />
                   <Route
